# src/main/webui/src/api/model/souscriptionSupplementaireDTO.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * sqq-inscription API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */
import type { CooperateurStatus } from "./cooperateurStatus";
import type { Instant } from "./instant";

export interface SouscriptionSupplementaireDTO {
  id?: number;
  prenom?: string;
  nom?: string;
  email?: string;
  partsSupplementaires?: number;
  status?: CooperateurStatus;
  createdAt?: Instant;
  updatedAt?: Instant;
}