# src/main/webui/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/administration/cooperateurs" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CooperateurDTO"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List",
        "tags" : [ "Admin Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/v1/administration/cooperateurs/{id}/process" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Mark As Processed",
        "tags" : [ "Admin Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/v1/administration/parts-additionnelles" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SouscriptionSupplementaireDTO"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List Souscriptions Supplementaires",
        "tags" : [ "Admin Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/v1/administration/parts-additionnelles/{id}/process" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Mark Souscription Supplementaire As Processed",
        "tags" : [ "Admin Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/v1/parts-supplementaires" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "prenom" : {
                    "type" : "string"
                  },
                  "nom" : {
                    "type" : "string"
                  },
                  "email" : {
                    "type" : "string"
                  },
                  "partsSupplementaires" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Register Form",
        "tags" : [ "Additional Shares Resource" ]
      }
    },
    "/api/v1/parts-supplementaires/success/{souscriptionId}" : {
      "post" : {
        "parameters" : [ {
          "name" : "souscriptionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Successful Payment",
        "tags" : [ "Additional Shares Resource" ]
      }
    },
    "/api/v1/registrations" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "genre" : {
                    "type" : "string"
                  },
                  "prenom" : {
                    "type" : "string"
                  },
                  "nom" : {
                    "type" : "string"
                  },
                  "telephone" : {
                    "type" : "string"
                  },
                  "email" : {
                    "type" : "string"
                  },
                  "adresse" : {
                    "type" : "string"
                  },
                  "ville" : {
                    "type" : "string"
                  },
                  "codePostal" : {
                    "type" : "string"
                  },
                  "etudiantOuMinimasSociaux" : {
                    "type" : "string"
                  },
                  "nombreDePersonnesDansLeFoyer" : {
                    "type" : "string"
                  },
                  "partsDeSoutien" : {
                    "type" : "string"
                  },
                  "acceptationDesStatus" : {
                    "type" : "string"
                  },
                  "binomeEnabled" : {
                    "type" : "string"
                  },
                  "binomeGenre" : {
                    "type" : "string"
                  },
                  "binomeNom" : {
                    "type" : "string"
                  },
                  "binomePrenom" : {
                    "type" : "string"
                  },
                  "binomeAdresse" : {
                    "type" : "string"
                  },
                  "binomeVille" : {
                    "type" : "string"
                  },
                  "binomeCodePostal" : {
                    "type" : "string"
                  },
                  "binomeEmail" : {
                    "type" : "string"
                  },
                  "binomeTelephone" : {
                    "type" : "string"
                  },
                  "binomeDateNaissance" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Register Form",
        "tags" : [ "Registration Resource" ]
      }
    },
    "/api/v1/registrations/success/{cooperateurId}" : {
      "post" : {
        "parameters" : [ {
          "name" : "cooperateurId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Successful Payment",
        "tags" : [ "Registration Resource" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "BinomeDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "genre" : {
            "$ref" : "#/components/schemas/Genre"
          },
          "nom" : {
            "type" : "string"
          },
          "prenom" : {
            "type" : "string"
          },
          "adresse" : {
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          },
          "codePostal" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "telephone" : {
            "type" : "string"
          },
          "dateNaissance" : {
            "type" : "string"
          }
        }
      },
      "CooperateurDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "genre" : {
            "$ref" : "#/components/schemas/Genre"
          },
          "prenom" : {
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "telephone" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "adresse" : {
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          },
          "codePostal" : {
            "type" : "string"
          },
          "etudiantOuMinimasSociaux" : {
            "type" : "boolean"
          },
          "nombreDePersonnesDansLeFoyer" : {
            "type" : "integer",
            "format" : "int64"
          },
          "parts" : {
            "type" : "integer",
            "format" : "int64"
          },
          "partsDeSoutien" : {
            "type" : "integer",
            "format" : "int64"
          },
          "acceptationDesStatus" : {
            "type" : "boolean"
          },
          "status" : {
            "$ref" : "#/components/schemas/CooperateurStatus"
          },
          "binome" : {
            "$ref" : "#/components/schemas/BinomeDTO"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "CooperateurStatus" : {
        "type" : "string",
        "enum" : [ "PAYMENT_PENDING", "PAID", "PROCESSED" ]
      },
      "Genre" : {
        "type" : "string",
        "enum" : [ "MADAME", "MONSIEUR" ]
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "SouscriptionSupplementaireDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "prenom" : {
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "partsSupplementaires" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "$ref" : "#/components/schemas/CooperateurStatus"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "scheme" : "basic",
        "description" : "Authentication"
      }
    }
  },
  "info" : {
    "title" : "sqq-inscription API",
    "version" : "1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}